Issue #28 - Move AwsS3Utils method used in test only to BaseInttest
